Prepare Dockerfile for production
The old standalone instance was created for testing purposes and was not
prepared to run in more complex distributed environments such as
Kubernetes.

This new deployment removes scripts and other components that were added
to the image in order to make it work in a standalone deployment. This
is still possible to use the old docker image with docker-compose but
only this new one is prepared to run in K8s.